refactor: Added class controller for AWS get presigned url
